Separate AppContainer as parent class for platform difference
